third_party/WebKit/LayoutTests/compositing/overflow/overflow-scroll-with-pointer-events-toggle-expected.txt

Verifies that toggling pointer-events correctly updates scrolling layers in compositor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS elementSubtreeHasCompositedScrollLayers(scroller) is false
PASS elementSubtreeHasCompositedScrollLayers(scroller) is true
PASS successfullyParsed is true

TEST COMPLETE

It should be possible to scroll this div.
